service/src/main/python/pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "yuca-client"
version = "0.1.0"
dependencies = ["psutil", "grpcio", "protobuf", "pynvml", "tqdm"]
